+++ html_support.nim
@@ -1,5 +1,5 @@
 import htmlparser, xmltree, strtabs, os, strutils, bb_system, streams,
-  globals_for_gh, macros
+  globals_for_gh, macros, lazy_rest_pkg/lrstgen
 
 ## gh_nimrod_doc_pages html support files.
 ##
@@ -15,6 +15,7 @@ import htmlparser, xmltree, strtabs, os, strutils, bb_system, streams,
 type
   First_tag = tuple[pos: int, tag: string]
   pair = tuple[src, dest: string] # Holds text replacement pairs.
+  Header_info* = tuple[level: int, href, text: string]
 
 
 proc post_process_html_local_links(html: PXmlNode, filename: string): bool =
@@ -182,6 +183,113 @@ proc post_process_html_local_links*(filename: string) =
 
     filename.write_file(buf)
     echo "Patching local links in ", filename
+
+
+proc recursive_text*(n: PXmlNode): string =
+  ## Extracts all possible inner text ignoring tags.
+  ##
+  ## The stdlib innerText proc doesn't deal with headers containing other inner
+  ## tags like ``<code>``.
+  result = ""
+  if n.kind != xnElement:
+    return
+
+  for child in n.items:
+    if child.kind in {xnText, xnEntity}:
+      result.add(child.text)
+    else:
+      result.add(child.recursive_text)
+
+
+proc extract_header(n: PXmlNode, result: var seq[Header_info]) =
+  ## Extracts from header node `n` the identifier and adds it to `result`.
+  ##
+  ## If the node can't be extracted safely `result` won't be changed. Empty
+  ## headers won't be added either.
+  assert n.kind == xnElement
+  if n.attrs.is_nil:
+    return
+
+  let text = n.recursive_text.strip
+  if text.len < 1:
+    return
+
+  var level = 0
+  case n.tag
+  of "h1": level = 1
+  of "h2": level = 2
+  of "h3": level = 3
+  of "h4": level = 4
+  of "h5": level = 5
+  of "h6": level = 6
+  else: return
+
+  if not n.attrs.has_key("id"):
+    return
+
+  result.add((level, n.attrs["id"], text))
+
+
+proc find_all_headers*(n: PXmlNode, result: var seq[Header_info]) =
+  ## Iterates over all the children of `n` returning those matching `h?`.
+  ##
+  ## Found headers will be appended to the `result` sequence, which can't be
+  ## nil or the proc will crash. The returned header level info is equal to the
+  ## tag level (``h1`` => 1).
+  assert result.not_nil
+  assert n.kind == xnElement
+
+  for child in n.items():
+    if child.kind != xnElement:
+      continue
+
+    case child.tag
+    of "h1", "h2", "h3", "h4", "h5", "h6":
+      child.extract_header(result)
+    else:
+      child.find_all_headers(result)
+
+
+proc find_all_headers(html: string): seq[Header_info] =
+  ## Wraps extraction of headers from an HTML file.
+  ##
+  ## Returns the found headers or the empty list.
+  let node = html.new_string_stream.parse_html
+  result = @[]
+  node.find_all_headers(result)
+
+
+proc tocify_markdown*(filename: string) =
+  ## Reads `filename` HTML and generates its index companion file.
+  ##
+  ## The `filename` usually contains the HTML of generated markdown code but in
+  ## theory this could work with anything. The generated index will be written
+  ## to the IdxExt variant.
+  ##
+  ## The format of the index file will conform to `Nim's idx file format
+  ## <http://nim-lang.org/docgen.html#index-idx-file-format>`_. If there is any
+  ## error or the index file ends up empty, the index file won't be created.
+  var toc = filename.read_file.find_all_headers
+  if toc.len < 1:
+    return
+
+  var
+    prefix = filename
+    GENERATOR: TRstGenerator
+
+  GENERATOR.init_rst_generator(out_html, filename)
+
+  if toc[0].level == 1:
+    # Special first entry reserved for the title.
+    GENERATOR.set_index_term("", toc[0].text)
+    system.delete(toc, 0)
+
+  for level, href, text in toc.items:
+    GENERATOR.set_index_term(href, text, level.repeat_char & text)
+
+  let idx = filename.change_file_ext("idx")
+  idx.remove_file
+  GENERATOR.write_index_file(idx)
 
 
 proc test() =